server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
fastapi==0.115.14
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.20.3
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.14
langchain-community==0.0.33
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.10.18
packaging==23.2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
